Split relational DDL execution
