+++ crates/truthdb-core/src/engine/relational/ddl/mod.rs
@@ -1,0 +1,11 @@
+mod alter;
+mod routines;
+mod security;
+mod table;
+mod views_indexes;
+
+pub(super) use alter::*;
+pub(super) use routines::*;
+pub(super) use security::*;
+pub(super) use table::*;
+pub(super) use views_indexes::*;
